requirements.txt
streamlit
streamlit-gsheets-connection
pandas
numpy
plotly